requirements.txt
langgraph
langgraph-prebuilt
langchain-core
langchain-community
langchain-openai
langgraph-cli
langchain
psycopg2-binary